Revert module rename
