+++ sdk/python/src/dagger/server/__init__.py
@@ -1,4 +1,4 @@
-from ._project import argument as argument
-from ._project import command as command
-from ._project import commands as commands
+from ._commands import argument as argument
+from ._commands import command as command
+from ._commands import commands as commands
 from ._server import Server as Server
